+++ docker/py3-native/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 ## 8.2.0
+* Updated py3-native to be based on ubi-9.1.
 * Updated the *LibreOffice* to the following version: **7.5.1**
 * Fixed an issue where sometimes when trying to retrieve *LibreOffice* installation from *download.documentfoundation.org* website, the chosen mirror was broken.
 * LibreOffice will now be installed always from the same mirror.